Browse, Swap pages up
Pages that allow for browsing furniture inventories and swapping the
location of a piece of furniture are up but not fully functional.
Need review on the change in item.js, not sure if I have added the link
to Bank table correctly.
Also please review users.js lines 28-30, which are an attempt to change
the location of the furniture item