Donate page added, can now add items
Donors can now submit an item to be donated- it will be added to the
items database but will not be linked to any furniture bank.
The Donate and Add Inventory pages can now write to the items database.

Hid the inactive links from the home page (to make it easier to
navigate?)